+++ pymunk/pyglet_util.py
@@ -52,7 +52,7 @@ class DrawOptions(pymunk.SpaceDebugDrawOptions):
         Typical usage::
 
         >>> import pymunk
-        >>> import pymunk.pygame_util
+        >>> import pymunk.pyglet_util
         >>> s = pymunk.Space()
         >>> options = pymunk.pyglet_util.DrawOptions()
         >>> s.debug_draw(options)